+++ packages/remix/lib/src/components/textfield/textarea_widget.dart
@@ -85,6 +85,4 @@ class RemixTextArea extends RemixTextField {
          expands: false,
          obscureText: false,
        );
-
-  static final styleFrom = TextFieldStyler.new;
 }